Pass ecl command line -foptions to eclcc and to deploy workunit
eclcc doesn't seem to use -foptions when creating an archive, but I will
pass them just in case.  Also passing to WUDeploy which does the actual
compilation.
